Persist multimedia data in last
