+++ src/promptflow-core/promptflow/storage/_queue_run_storage.py
@@ -3,10 +3,7 @@
 # ---------------------------------------------------------
 
 from multiprocessing import Queue
-from pathlib import Path
 
-from promptflow._constants import OutputsFolderName
-from promptflow._utils.multimedia_utils import MultimediaProcessor
 from promptflow.contracts.run_info import FlowRunInfo
 from promptflow.contracts.run_info import RunInfo as NodeRunInfo
 from promptflow.storage import AbstractRunStorage
@@ -25,25 +22,3 @@ class QueueRunStorage(AbstractRunStorage):
 
     def persist_flow_run(self, run_info: FlowRunInfo):
         self.queue.put(run_info)
-
-
-class ServiceQueueRunStorage(QueueRunStorage):
-    """This storage persist multimedia data after run info is put into the output queue."""
-
-    def __init__(self, queue: Queue, output_dir: Path):
-        super().__init__(queue)
-        self._flow_outputs_path = output_dir / OutputsFolderName.FLOW_OUTPUTS
-        self._flow_artifacts_path = output_dir / OutputsFolderName.FLOW_ARTIFACTS
-        self._node_artifacts_path = output_dir / OutputsFolderName.NODE_ARTIFACTS
-
-    def persist_node_run(self, run_info: NodeRunInfo):
-        super().persist_node_run(run_info)
-        node_folder = self._node_artifacts_path / str(run_info.index) / run_info.node
-        multimedia_processor = MultimediaProcessor.create(run_info.message_format)
-        multimedia_processor.process_multimedia_in_run_info(run_info, node_folder)
-
-    def persist_flow_run(self, run_info: FlowRunInfo):
-        super().persist_flow_run(run_info)
-        flow_folder = self._flow_artifacts_path / str(run_info.index)
-        multimedia_processor = MultimediaProcessor.create(run_info.message_format)
-        multimedia_processor.process_multimedia_in_run_info(run_info, flow_folder)
